pyproject.toml
[project]
name = "self-rag-legal"
version = "0.1.0"
description = "Self-RAG System for Legal Document Analysis with Hallucination Mitigation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Academic Use Only"}
authors = [
    {name = "DSC261 Project Team"},
]

dependencies = [
    # Core ML and NLP
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "sentence-transformers>=2.2.2",
    "datasets>=2.14.0",

    # Efficient training and quantization
    "peft>=0.7.0",
    "bitsandbytes>=0.41.0",
    "accelerate>=0.24.0",
    "trl>=0.7.0",

    # Retrieval and vector databases
    "faiss-cpu>=1.7.4",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",

    # LLM inference optimization
    "llama-cpp-python>=0.2.0",

    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",

    # Evaluation and metrics
    "rouge-score>=0.1.2",
    "nltk>=3.8.1",
    "sacrebleu>=2.3.1",

    # Utilities
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",

    # Jupyter and visualization
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",

    # API clients (optional)
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
